requirements.txt
requests>=2.31.0
sqlalchemy==1.4.54
pandas==2.2.3
psycopg2-binary==2.9.10
streamlit>=1.35.0
python-dotenv>=1.0.1
pytest==8.3.5
pytest-mock==3.14.0
pytest-cov==6.0.0
streamlit==1.41.1
plotly==5.24.1